backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
MetaTrader5>=5.0.45
pandas>=2.2.0
python-dotenv>=1.0.0
openai>=1.30.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
cryptography>=41.0.0
bcrypt>=4.0.0
httpx>=0.27.0
boto3>=1.35.0
python-multipart>=0.0.9
telethon>=1.34.0
celery[redis]>=5.3.0
redis>=5.0.0
openpyxl>=3.1.0
instructor>=1.3.0
alembic>=1.13.0
slowapi>=0.1.9
pytest>=8.0.0
pytest-asyncio>=0.23.0
python-json-logger>=2.0.7
limits[redis]>=3.13.0
sentry-sdk[fastapi]>=2.0.0
flower>=2.0.0
